Put node config back in the node
