+++ Cargo.toml
@@ -110,6 +110,7 @@ chrono = { version = "0.4", features = ["serde"] }
 hex = { version = "0.4.3" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1" }
+serde_with = "3.12.0"
 sha3 = "0.10.8"
 strum_macros = "0.27.1"
 tracing = "0.1"
