+++ crates/hyle-modules/Cargo.toml
@@ -20,7 +20,6 @@ serde = "1.0"
 serde_json = { version = "1" }
 tracing = "0.1"
 strum_macros = "0.27.1"
-serde_with = "3.12.0"
 config = { version = "0.15.9", default-features = false, features = ["toml"] }
 assertables = { version = "9.5.2" }
 anymap = { version = "0.12.1" }
