+++ crates/hyle-modules/src/utils/mod.rs
@@ -1,4 +1,3 @@
-pub mod conf;
 pub mod da_codec;
 pub mod logger;
 pub mod profiling;
