+++ src/model.rs
@@ -1,7 +1,7 @@
 //! Various data structures
 
 use hyle_crypto::SharedBlstCrypto;
-use hyle_modules::{modules::SharedBuildApiCtx, utils::conf::SharedConf};
+use hyle_modules::modules::SharedBuildApiCtx;
 
 // Re-export
 pub use hyle_model::*;
@@ -9,6 +9,8 @@ pub use hyle_model::*;
 mod indexer;
 
 pub use indexer::*;
+
+use crate::utils::conf::SharedConf;
 
 #[derive(Clone)]
 pub struct SharedRunContext {
